student/К3341/Саид_Наваф/Lr3/requirements.txt
Django>=4.2,<5
djangorestframework
djoser
djangorestframework-simplejwt
drf-spectacular
mkdocs
mkdocs-material
django-cors-headers